+++ src/infinifix/__init__.py
@@ -1,3 +1,3 @@
 """InfiniFix - Huawei Linux Doctor."""
 
-__version__ = "0.2.0b1"
+__version__ = "0.3.0b1"